ZCS-11657: Create New SOAP API for testing S3 bucket config
Revert "Create New SOAP API for testing S3 bucket config"

This reverts commit a016e5814ecb0ff61cb1fbf920dd490571df2c6e.

Revert "Create New SOAP API for testing S3 bucket config"

This reverts commit 07c8d30c74c8494d36e16e1f6938c740f683e872.

added new api to validate S3 bucket
