--- common/src/java/com/zimbra/common/soap/GlobalExternalStoreConfigConstants.java
@@ -1,17 +1,0 @@
-package com.zimbra.common.soap;
-
-public final class GlobalExternalStoreConfigConstants {
-
-    public static final String E_GLOBAL_S3_BUCKET_CONFIGURATIONS = "global/s3BucketConfigurations";
-    public static final String A_S3_GLOBAL_BUCKET_UUID = "globalBucketUUID";
-    public static final String A_S3_BUCKET_NAME = "bucketName";
-    public static final String A_S3_STORE_PROVIDER = "storeProvider";
-    public static final String A_S3_PROTOCOL = "protocol";
-    public static final String A_S3_ACCESS_KEY = "accessKey";
-    public static final String A_S3_SECRET_KEY = "secretKey";
-    public static final String A_S3_REGION = "region";
-    public static final String A_S3_DESTINATION_PATH = "destinationPath";
-    public static final String A_S3_URL = "url";
-    public static final String A_S3_BUCKET_STATUS = "bucketStatus";
-
-}
