--- soap/src/java/com/zimbra/soap/admin/message/CreateS3BucketConfigRequest.java
@@ -1,39 +1,0 @@
-/*
- * ***** BEGIN LICENSE BLOCK *****
- * Zimbra Collaboration Suite Server
- * Copyright (C) 2022 Synacor, Inc.
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software Foundation,
- * version 2 of the License.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU General Public License for more details.
- * You should have received a copy of the GNU General Public License along with this program.
- * If not, see <https://www.gnu.org/licenses/>.
- * ***** END LICENSE BLOCK *****
- */
-
-package com.zimbra.soap.admin.message;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
-
-import com.zimbra.common.soap.AdminConstants;
-import com.zimbra.soap.admin.type.AdminAttrsImpl;
-
-/**
- * @zm-api-command-auth-required true
- * @zm-api-command-admin-auth-required true
- * @zm-api-command-description Create S3 Bucket Config <br />
- */
-@XmlAccessorType(XmlAccessType.NONE)
-@XmlRootElement(name = AdminConstants.E_CREATE_S3_BUCKET_CONFIG_REQUEST)
-public class CreateS3BucketConfigRequest extends AdminAttrsImpl {
-
-    public CreateS3BucketConfigRequest() {
-    }
-
-}
